+++ cmd/login/flag.go
@@ -12,10 +12,11 @@ const (
 	flagInternalAPI    = "internal-api"
 	callbackServerPort = "callback-port"
 
-	flagWCName         = "workload-cluster"
-	flagWCOrganization = "organization"
-	flagWCCertGroups   = "certificate-group"
-	flagWCCertTTL      = "certificate-ttl"
+	flagWCName          = "workload-cluster"
+	flagWCOrganization  = "organization"
+	flagWCCertGroups    = "certificate-group"
+	flagWCCertTTL       = "certificate-ttl"
+	flagWCSelfContained = "self-contained"
 )
 
 type flag struct {
@@ -23,10 +24,11 @@ type flag struct {
 	ClusterAdmin       bool
 	InternalAPI        bool
 
-	WCName         string
-	WCOrganization string
-	WCCertGroups   []string
-	WCCertTTL      string
+	WCName          string
+	WCOrganization  string
+	WCCertGroups    []string
+	WCCertTTL       string
+	WCSelfContained string
 
 	config genericclioptions.RESTClientGetter
 }
@@ -40,6 +42,7 @@ func (f *flag) Init(cmd *cobra.Command) {
 	cmd.Flags().StringVar(&f.WCOrganization, flagWCOrganization, "", fmt.Sprintf("Organization that owns the workload cluster. Requires --%s.", flagWCName))
 	cmd.Flags().StringSliceVar(&f.WCCertGroups, flagWCCertGroups, nil, fmt.Sprintf("RBAC group name to be encoded into the X.509 field \"O\". Requires --%s.", flagWCName))
 	cmd.Flags().StringVar(&f.WCCertTTL, flagWCCertTTL, "1h", fmt.Sprintf("How long the client certificate should live for. Requires --%s.", flagWCName))
+	cmd.Flags().StringVar(&f.WCSelfContained, flagWCSelfContained, "", fmt.Sprintf("Create a self-contained kubectl config with embedded credentials and write it to this path. Requires --%s.", flagWCName))
 
 	f.config = genericclioptions.NewConfigFlags(true)
 	f.config.(*genericclioptions.ConfigFlags).AddFlags(cmd.Flags())